@@ -1,11 +1,11 @@
 name: flutter_icons
 description: Customizable Icons for Flutter，you can use with over 3K+ icons in your flutter project
-version: 1.1.0
+version: 1.1.1-nullsafety.0
 author: flutter-studio<2534290808@qq.com>
 homepage: https://github.com/flutter-studio/flutter-icons.git
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
@@ -21,54 +21,54 @@ dev_dependencies:
 # The following section is specific to Flutter.
 flutter:
   fonts:
-      - family: Ionicons
-        fonts:
-          - asset: fonts/Ionicons.ttf
-      - family: AntDesign
-        fonts:
-          - asset: fonts/AntDesign.ttf
-      - family: FontAwesome
-        fonts:
-          - asset: fonts/FontAwesome.ttf
-      - family: MaterialIcons
-        fonts:
-          - asset: fonts/MaterialIcons.ttf
-      - family: Entypo
-        fonts:
-          - asset: fonts/Entypo.ttf
-      - family: EvilIcons
-        fonts:
-          - asset: fonts/EvilIcons.ttf
-      - family: Feather
-        fonts:
-          - asset: fonts/Feather.ttf
-      - family: Foundation
-        fonts:
-          - asset: fonts/Foundation.ttf
-      - family: MaterialCommunityIcons
-        fonts:
-          - asset: fonts/MaterialCommunityIcons.ttf
-      - family: Octicons
-        fonts:
-          - asset: fonts/Octicons.ttf
-      - family: SimpleLineIcons
-        fonts:
-          - asset: fonts/SimpleLineIcons.ttf
-      - family: Zocial
-        fonts:
-          - asset: fonts/Zocial.ttf
-      - family: FontAwesome5
-        fonts:
-          - asset: fonts/FontAwesome5_Regular.ttf
-      - family: FontAwesome5_Brands
-        fonts:
-          - asset: fonts/FontAwesome5_Brands.ttf
-      - family: FontAwesome5_Solid
-        fonts:
-          - asset: fonts/FontAwesome5_Solid.ttf
-      - family: WeatherIcons
-        fonts:
-          - asset: fonts/weathericons.ttf
+    - family: Ionicons
+      fonts:
+        - asset: fonts/Ionicons.ttf
+    - family: AntDesign
+      fonts:
+        - asset: fonts/AntDesign.ttf
+    - family: FontAwesome
+      fonts:
+        - asset: fonts/FontAwesome.ttf
+    - family: MaterialIcons
+      fonts:
+        - asset: fonts/MaterialIcons.ttf
+    - family: Entypo
+      fonts:
+        - asset: fonts/Entypo.ttf
+    - family: EvilIcons
+      fonts:
+        - asset: fonts/EvilIcons.ttf
+    - family: Feather
+      fonts:
+        - asset: fonts/Feather.ttf
+    - family: Foundation
+      fonts:
+        - asset: fonts/Foundation.ttf
+    - family: MaterialCommunityIcons
+      fonts:
+        - asset: fonts/MaterialCommunityIcons.ttf
+    - family: Octicons
+      fonts:
+        - asset: fonts/Octicons.ttf
+    - family: SimpleLineIcons
+      fonts:
+        - asset: fonts/SimpleLineIcons.ttf
+    - family: Zocial
+      fonts:
+        - asset: fonts/Zocial.ttf
+    - family: FontAwesome5
+      fonts:
+        - asset: fonts/FontAwesome5_Regular.ttf
+    - family: FontAwesome5_Brands
+      fonts:
+        - asset: fonts/FontAwesome5_Brands.ttf
+    - family: FontAwesome5_Solid
+      fonts:
+        - asset: fonts/FontAwesome5_Solid.ttf
+    - family: WeatherIcons
+      fonts:
+        - asset: fonts/weathericons.ttf
 
   # To add assets to your package, add an assets section, like this:
   # assets:
